pyproject.toml
[tool.poetry]
name = "damn-tool"
version = "0.1.4"
description = "The DAMN (Data Assets Metric Navigation) tool extracts and reports metrics about your data assets"
authors = ["Olivier Dupuis <olivier@republicofdata.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
termcolor = "^2.3.0"
pyyaml = "^6.0"
jinja2 = "^3.1.2"
boto3 = "^1.28.1"
pyperclip = "^1.8.2"
line-profiler = "^4.0.3"
snowflake-connector-python = "^3.0.4"
google-cloud-bigquery = "^3.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
damn = "damn_tool.__init__:cli"
